upgrade questdb 6.7 remove minio (wanted it committed for future use)

+++ src/main/java/io/quest/metadata/CounterFileReader.java
@@ -26,7 +26,7 @@ public class CounterFileReader implements Closeable {
 
     private final FilesFacade ff;
 
-    private long uniqueIdFd = -1;
+    private int uniqueIdFd = -1;
     private long uniqueIdMem = 0;
 
     public CounterFileReader(FilesFacade ff) {
